refactor: centralize category icons
