--- Frontend-nextjs/lib/icons.ts
@@ -1,161 +1,0 @@
-// icons.ts
-import {
-    FiTag,
-    FiHome,
-    FiTrendingUp,
-    FiGift,
-    FiBookOpen,
-    FiBriefcase,
-    FiShoppingCart,
-    FiCreditCard,
-    FiCoffee,
-    FiHeart,
-    FiUser,
-    FiTruck,
-    FiCalendar,
-    FiGlobe,
-    FiStar,
-    FiCamera,
-    FiPhone,
-    FiDroplet,
-    FiBell,
-    FiSun,
-    FiMoon,
-} from "react-icons/fi";
-import {
-    FaUtensils,
-    FaCar,
-    FaRegSmile,
-    FaHeartbeat,
-    FaPlane,
-    FaPiggyBank,
-    FaRegMoneyBillAlt,
-    FaBus,
-    FaBolt,
-    FaTree,
-    FaDumbbell,
-    FaBicycle,
-    FaLaptop,
-    FaPaw,
-    FaRegSun,
-    FaBeer,
-    FaMedkit,
-    FaCameraRetro,
-    FaWallet,
-    FaShoppingBasket,
-    FaGlasses,
-    FaLeaf,
-    FaBook,
-    FaTshirt,
-    FaDog,
-    FaCat,
-} from "react-icons/fa";
-import {
-    MdAttachMoney,
-    MdHealthAndSafety,
-    MdPets,
-    MdWork,
-    MdLocalBar,
-    MdDirectionsCar,
-    MdRestaurant,
-    MdFlight,
-    MdHome,
-    MdSportsSoccer,
-    MdSchool,
-    MdChildCare,
-    MdDirectionsBike,
-    MdShoppingBag,
-    MdLocalGroceryStore,
-    MdSavings,
-    MdChair,
-    MdFastfood,
-    MdLocalCafe,
-    MdLocalHospital,
-    MdSpa,
-    MdBeachAccess,
-    MdOutlineSportsHandball,
-    MdOutlineTrain,
-    MdOutlineFlightTakeoff,
-    MdOutlineShoppingCart,
-} from "react-icons/md";
-
-export const CATEGORY_ICONS_MAP = {
-    // Feather
-    FiTag,
-    FiHome,
-    FiTrendingUp,
-    FiGift,
-    FiBookOpen,
-    FiBriefcase,
-    FiShoppingCart,
-    FiCreditCard,
-    FiCoffee,
-    FiHeart,
-    FiUser,
-    FiTruck,
-    FiCalendar,
-    FiGlobe,
-    FiStar,
-    FiCamera,
-    FiPhone,
-    FiDroplet,
-    FiBell,
-    FiSun,
-    FiMoon,
-
-    // FontAwesome
-    FaUtensils,
-    FaCar,
-    FaRegSmile,
-    FaHeartbeat,
-    FaPlane,
-    FaPiggyBank,
-    FaRegMoneyBillAlt,
-    FaBus,
-    FaBolt,
-    FaTree,
-    FaDumbbell,
-    FaBicycle,
-    FaLaptop,
-    FaPaw,
-    FaRegSun,
-    FaBeer,
-    FaMedkit,
-    FaCameraRetro,
-    FaWallet,
-    FaShoppingBasket,
-    FaGlasses,
-    FaLeaf,
-    FaBook,
-    FaTshirt,
-    FaDog,
-    FaCat,
-
-    // Material
-    MdAttachMoney,
-    MdHealthAndSafety,
-    MdPets,
-    MdWork,
-    MdLocalBar,
-    MdDirectionsCar,
-    MdRestaurant,
-    MdFlight,
-    MdHome,
-    MdSportsSoccer,
-    MdSchool,
-    MdChildCare,
-    MdDirectionsBike,
-    MdShoppingBag,
-    MdLocalGroceryStore,
-    MdSavings,
-    MdChair,
-    MdFastfood,
-    MdLocalCafe,
-    MdLocalHospital,
-    MdSpa,
-    MdBeachAccess,
-    MdOutlineSportsHandball,
-    MdOutlineTrain,
-    MdOutlineFlightTakeoff,
-    MdOutlineShoppingCart,
-};
