+++ Frontend-nextjs/utils/categoryOptions.ts
@@ -1,3 +1,84 @@
+import type { IconType } from "react-icons";
+import {
+    FiTag,
+    FiHome,
+    FiTrendingUp,
+    FiGift,
+    FiBookOpen,
+    FiBriefcase,
+    FiShoppingCart,
+    FiCreditCard,
+    FiCoffee,
+    FiHeart,
+    FiUser,
+    FiTruck,
+    FiCalendar,
+    FiGlobe,
+    FiStar,
+    FiCamera,
+    FiPhone,
+    FiDroplet,
+    FiBell,
+    FiSun,
+    FiMoon,
+} from "react-icons/fi";
+import {
+    FaUtensils,
+    FaCar,
+    FaRegSmile,
+    FaHeartbeat,
+    FaPlane,
+    FaPiggyBank,
+    FaRegMoneyBillAlt,
+    FaBus,
+    FaBolt,
+    FaTree,
+    FaDumbbell,
+    FaBicycle,
+    FaLaptop,
+    FaPaw,
+    FaRegSun,
+    FaBeer,
+    FaMedkit,
+    FaCameraRetro,
+    FaWallet,
+    FaShoppingBasket,
+    FaGlasses,
+    FaLeaf,
+    FaBook,
+    FaTshirt,
+    FaDog,
+    FaCat,
+} from "react-icons/fa";
+import {
+    MdAttachMoney,
+    MdHealthAndSafety,
+    MdPets,
+    MdWork,
+    MdLocalBar,
+    MdDirectionsCar,
+    MdRestaurant,
+    MdFlight,
+    MdHome,
+    MdSportsSoccer,
+    MdSchool,
+    MdChildCare,
+    MdDirectionsBike,
+    MdShoppingBag,
+    MdLocalGroceryStore,
+    MdSavings,
+    MdChair,
+    MdFastfood,
+    MdLocalCafe,
+    MdLocalHospital,
+    MdSpa,
+    MdBeachAccess,
+    MdOutlineSportsHandball,
+    MdOutlineTrain,
+    MdOutlineFlightTakeoff,
+    MdOutlineShoppingCart,
+} from "react-icons/md";
+
 // =======================
 // PALETTE COLORI ESTESA (MAXI, con nome)
 // =======================
@@ -83,7 +164,7 @@ export const CATEGORY_COLORS = [
     { value: "#cbd5e1", name: "Silver" }, // tailwind slate-100
     { value: "#fff", name: "Bianco (sconsigliato)" },
     { value: "#1c1917", name: "Quasi Nero (sconsigliato)" },
-];
+] as const;
 
 // ===============================
 // ICONS ESTESA (Feather, FontAwesome, Material...)
@@ -167,4 +248,89 @@ export const CATEGORY_ICONS = [
     { value: "MdOutlineTrain", name: "Treno" },
     { value: "MdOutlineFlightTakeoff", name: "Decollo" },
     { value: "MdOutlineShoppingCart", name: "Carrello (Alt)" },
-];
+] as const;
+
+export const CATEGORY_ICONS_MAP = {
+    FiTag,
+    FiHome,
+    FiTrendingUp,
+    FiGift,
+    FiBookOpen,
+    FiBriefcase,
+    FiShoppingCart,
+    FiCreditCard,
+    FiCoffee,
+    FiHeart,
+    FiUser,
+    FiTruck,
+    FiCalendar,
+    FiGlobe,
+    FiStar,
+    FiCamera,
+    FiPhone,
+    FiDroplet,
+    FiBell,
+    FiSun,
+    FiMoon,
+
+    FaUtensils,
+    FaCar,
+    FaRegSmile,
+    FaHeartbeat,
+    FaPlane,
+    FaPiggyBank,
+    FaRegMoneyBillAlt,
+    FaBus,
+    FaBolt,
+    FaTree,
+    FaDumbbell,
+    FaBicycle,
+    FaLaptop,
+    FaPaw,
+    FaRegSun,
+    FaBeer,
+    FaMedkit,
+    FaCameraRetro,
+    FaWallet,
+    FaShoppingBasket,
+    FaGlasses,
+    FaLeaf,
+    FaBook,
+    FaTshirt,
+    FaDog,
+    FaCat,
+
+    MdAttachMoney,
+    MdHealthAndSafety,
+    MdPets,
+    MdWork,
+    MdLocalBar,
+    MdDirectionsCar,
+    MdRestaurant,
+    MdFlight,
+    MdHome,
+    MdSportsSoccer,
+    MdSchool,
+    MdChildCare,
+    MdDirectionsBike,
+    MdShoppingBag,
+    MdLocalGroceryStore,
+    MdSavings,
+    MdChair,
+    MdFastfood,
+    MdLocalCafe,
+    MdLocalHospital,
+    MdSpa,
+    MdBeachAccess,
+    MdOutlineSportsHandball,
+    MdOutlineTrain,
+    MdOutlineFlightTakeoff,
+    MdOutlineShoppingCart,
+} as const;
+
+export type CategoryIconName = keyof typeof CATEGORY_ICONS_MAP;
+
+export function getIconComponent(name?: CategoryIconName): IconType {
+    if (name && CATEGORY_ICONS_MAP[name]) return CATEGORY_ICONS_MAP[name];
+    return FiTag;
+}
